Rename: HttpMethod enum 객체 util 패키지로 이동
